+++ example/lib/main.dart
@@ -43,6 +43,11 @@ class _ExampleState extends State<Example> {
                   onTap: () => setState(() => _amount += 2500),
                   child: TextMorph('$_amount', morph: const MorphStyle.roll()),
                 ),
+                const SizedBox(height: 16),
+                GestureDetector(
+                  onTap: () => setState(() => _amount += 2500),
+                  child: TickerText('$_amount'),
+                ),
               ],
             ),
           ),